CarriageRegister and PackageRegister components completed
