+++ src/App.vue
@@ -1,54 +1,20 @@
 <script setup lang="ts">
+import {onBeforeMount, ref} from "vue";
 //@ts-ignore
-import Stepper from "@/components/core/Stepper.vue";
-//@ts-ignore
-import type  { Steps, StepComponent} from "@/types/intefaces";
-//@ts-ignore
-import LocationOrigin from "@/components/packages/stepper/LocationOrigin.vue";
-//@ts-ignore
-import LocationDestiny from "@/components/packages/stepper/LocationDestiny.vue";
-//@ts-ignore
-import PackageDimensions from "@/components/packages/stepper/PackageDimensions.vue";
-//@ts-ignore
-import PackageWeight from "@/components/packages/stepper/PackageWeight.vue";
-//@ts-ignore
-import Ticket from "@/components/packages/stepper/Ticket.vue";
-//@ts-ignore
-import Success from "@/components/packages/stepper/Success.vue";
 import CarriageRegister from "@/components/carriages/stepper/CarriageRegister.vue";
+//@ts-ignore
+import PackageRegister from "@/components/packages/stepper/PackageRegister.vue";
 
-const firstStep: StepComponent = {
-  finalStep: false,
-  value: LocationOrigin,
-}
-const secondtStep: StepComponent = {
-  finalStep: false,
-  value: LocationDestiny
-}
-const thirdStep: StepComponent = {
-  finalStep: false,
-  value: PackageDimensions
-}
-const fourthStep: StepComponent = {
-  finalStep: false,
-  value: PackageWeight
-}
-const fifthStep: StepComponent = {
-  finalStep: true,
-  value: Ticket
-}
-const finalStep: StepComponent = {
-  finalStep: false,
-  value: Success
-}
-const appStep: Steps = {
-  steps: 6,
-  listStepsComponents: [firstStep, secondtStep, thirdStep, fourthStep, fifthStep, finalStep]
-}
+const random = ref<number>(0);
+
+onBeforeMount(() => {
+  random.value = Math.random() ;
+})
 </script>
 
 <template>
-  <CarriageRegister />
+ <CarriageRegister v-if="random<0.51"/>
+ <PackageRegister v-else/>
 </template>
 
 <style scoped>
